+++ releasenotes.md
@@ -5,6 +5,7 @@
 - add support on key is a number
 - added getResourceBundle to API
 - allow multiple post-processors
+- fallback to singular if no plural is found fixes issue [issue 356](https://github.com/i18next/i18next/issues/356) 
 
 ### 1.7.7
 - fixes issue with stack overflow on t(lng, count)